Create main.py
Main script to scrape the autoscout.
In the main function, one needs to set the manufacture, model and year of the car as well as the post code it needs to look into it.
To change the country, the main link to  the website needs to be changed.
The results of all the advertisements are saved in a csv file.